sensor datalarının uart üzerinden gönderilmesi sağlandı.

+++ Projects/WL55E/Applications/SubGHz_Phy/SubGHz_Phy_PingPong/Sensor/Inc/ltr390uv.h
@@ -153,6 +153,6 @@ int ltr390uv_init(void);
 
 int ltr390uv_read_id(void);
 
-void ltr390uv_read_uvs_data(float* data);
+void ltr390uv_read_uvs_data(uint32_t* data);
 
 #endif
